requirements.txt
# Core dependencies
numpy>=1.21.0
faiss-cpu>=1.7.0  # Required for vector similarity search
groq>=0.3.0  # Required for Groq model integration
python-dotenv>=0.19.0
loguru>=0.5.3

# LangChain dependencies (required for model integration)
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0

# CLI dependencies
click>=8.0.0
rich>=10.0.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
mypy>=0.900 